source-test/requirements.txt
flask
mysql-connector-python
prometheus_client
tenacity
